refactor: 카드 등록 관련 화면을 별도 페이지로 분리
